Clarify how to use set_dark/light_theme()
Because it said "No need to include `library(ThemeSwitcher)` in your R scripts." But I couldn't run set_light_theme() until the package was loaded, which was a bit confusing. The change makes it a bit clearer how to run the set_dark/light_theme() functions. It's a great addin!